Improve overload chain docstrings for Sphinx
Sphinx can nowadays nicely parse overload chains if all function
signatures are printed at the beginning of the docstring. This change
therefore moves signatures to the beginning to improve Sphinx output.

When an overload chain specifies multiple overload-specific docstrings,
another copy of the signature (in ``code quotes``) is inserted later on
for context.

This change resembles a PR in the pybind11 repository
(https://github.com/pybind/pybind11/issues/2619).