Refactored nanobind so that it works with Py_LIMITED_API
There is an ongoing effort to refactor CPython internals to improve its
performance. This has serious consequences for tools like nanobind,
which rely on various CPython implementation details that are now
subject to change.

This commit changes nanobind so that it can (optionally) work via the
Py_LIMITED_API, which means that it treats all CPython data structures
as fully opaque and only accesses them through an official API/ABI with
long-term stability.

This requires a change that is pending for inclusion into Python 3.11
(issue 39012).