fixed function-level documentation in pydoc
commit 0b3548ad6faf4e changed function objects in a subtle way that broke
docstring extraction in pydoc. While the nb_func.__doc__ member worked
as expected, pydoc no longer queried it.

This commit both fixes the regression and adds tests to ensure that it
doesn't happen again in the future (by running pydoc and looking for
certain strings that we would expect to be part of the result).