Exploit move constructors in def_readwrite and def_readwrite_static
This commit implements a simple optimization that applies when the
following two conditions hold:

- A field in a bound C++ types is exposed as writeable Python property

- Setting the field involves an inter-language type conversion step
  (e.g., for `type_caster<std::vector<T>>` implemented in
  ``nanobind/stl/vector.h``)

In this case, we can avoid a superfluous copying step by
move-constructing the constructed instance from the type caster directly
into the field.

This commit also adds tests to verify that this actually happens under
the hood.