detail::func_create() tweak that leads to a binary size reduction
On Clang (tested on macOS, but other platforms are likely affected as
well), adding an ``__attribute__((always_inline))`` to the main
anonymous function (``impl``) in ``nanobind::detail::func_create()`` has
a noticeable impact on the quality of the generated code.

In the ``func_nanobind_opt`` benchmark, binary size goes down by ~5%.

It's a little odd that this matters, because the function can in any
case not be inlined (we store a function pointer to it in a data
structure managed by the nanobind library). It's likely that the
attribute also enables more favorable code generation heuristics for the
body of the function, hence the choice to merge this into the main
branch.