Tweaks to type annotations in docstrings, STL dict casters
Besides minor syntactic changes, this commit:

1. Generalizes casters of ``std::unordered_map<..>`` and
   ``std::map<..>`` so that they can load data from *any* abstract
   mapping and not just concrete ``dict`` instances.

2. Changes the casters of ``std::vector<..>``, ``std::list<..>``, and
   ``std::array<..>`` to produce a concrete ``list[..]`` type annotation
   in function arguments/return values instead of the previously used
   abstract ``Sequence[..]`` annotation.

Both cases now lead to slight disagreements between type annotations in
docstrings and the actual set of types that is accepted as a valid
input. I find such slight imperfections in type annotations acceptable,
when fixing them would introduce other undesirable aspects.
In this case:

1. Readability: compare, e.g., ``set[str]`` to the theoretically correct
   ``typing.AbstractSet[str]``. I strongly prefer the former.

2. Type caster type annotations would need to be different for
   function arguments (abstract) vs. return values (concrete), which
   introduces quite a bit of extra complexity into nanobind.

Hence, from now on the general principle is that the nanobind type
casters are a going to be little bit more accommodating than what their
signatures strictly specify, which in turn leads to more readable
function signatures and keeps the codebase simple.

For those that need a signature to be ultra-correct, there is still
nb::raw_doc() as an escape hatch.