Vector call dispatch loop optimizations
Nanobind provides a "simple" and a "complex" vector call dispatch
routine. The complex variant is required for functions with default or
keyword arguments, while this simple variant can ignore those aspects
and has lower overheads.

This commit further specializes the simple variant to functions with 8
or less arguments. This makes it possible to remove a call to
``alloca()`` (which impedes optimizations) and a ``memset()`` with an
unknown size (which cannot be optimized by the compiler). Apart from
this change, the commit also adds code layout hints for GCC and Clang.

Following this change, I see speedups of 4-5% for function and method
call call microbenchmarks.