Preliminary adaptations for PyPy
This commit lays the foundation for eventually supporting PyPy as an
alternative interpreter besides CPython. This is is dependent on a few
remaining issues being worked out on the PyPy end:

- https://foss.heptapod.net/pypy/pypy/-/issues/3847
- https://foss.heptapod.net/pypy/pypy/-/issues/3845
- https://foss.heptapod.net/pypy/pypy/-/issues/3844

I decided to already merge this PR because the changes are also nice to
have just for the CPython part (in particular, the alternative local
implementation of `PyType_FromMetaclass` is more general/robust).