Fix dangling tp_members pointer.
In a heap-allocated type, tp_members normally points to the interior of
the type object, at offset tp_basicsize. By copying the tp_members field
of a temporary heap-allocated type object to the final type object, we
leave a dangling pointer. Instead, allocate space for the PyMemberDef
objects at the end of the nanobind-allocated type object, copy them into
place, and set tp_members to point to that location.