Allow nanobind methods on non-nanobind classes.
This allows us to extend an existing Python (or
pybind11!) class with methods defined in nanobind. This is handy if you
want to add a small amount of nanobind code to a class
defined in other ways. For example, one use is to migrate methods of a
pybind11-defined class to nanobind one by one.

All we have to do to allow this is remove the type check. I
can't see anything that goes wrong if we simply allow dispatch to
proceed.