More careful check to see if a type is copy-constructible
The design of `std::is_copy_constructible` and is somewhat broken in
that this functions returns `true` whenever a copy constructor exists,
even if it cannot actually be compiled. This happens, e.g., for
`std::vector<T>` when `T` is not copy-constructible.

This commit ports a custom type trait from pybind11 to address this
issue. By default, it is simply a wrapper around the original
`std::is_copy_constructible<>`.

The added file `include/nanobind/stl/detail/traits.h` adds partial
overloads for STL types and should be included by a future
`bind_vector.h` and `bind_map.h` implementation.