Feature to customize argument type signatures
This commit adds an API for overriding a function's type annotation in
generated docstrings. It is particularly helpful when the type signature
is not obvious and must be computed at compile time. Otherwise, the
`nb::raw_doc()` attribute provides a simpler alternative for taking full
control of docstrings and type annotations.

Consider the following binding that iterates over a Python list.

```cpp
m.def("f", [](nb::list l) {
    for (handle h : l) {
        // ...
    }
});
```

Suppose that `f` expects a list of `Foo` objects, which is not clear
from the signature. To improve the function signature, use the
`nb::typed<T, D>` wrapper class to pass the argument. This wrapper is
defined as follows:

```cpp
template <typename T, typename D> struct typed {
    T value;
};
```

Here, `T` refers to the original argument type, and `D` points to a
helper class that will be used to compute the type name at runtime.

This means that any access to the list `l` must be replaced by
`l.value`:

```cpp
m.def("f", [](nb::typed<nb::list, FooListName> l) {
    for (nb::handle h : l.value) {
        // ...
    }
});
```

In this simple example, the `FooListName` type can be defined as follows:

```cpp
struct FooListName {
    static constexpr auto Name =
        nb::detail::const_name("list[") +
        nb::detail::const_name<Foo>() +
        nb::detail::const_name("]");
};
```

More generally, this can be a templated class with partial overloads,
which allows for advanced constructions.