Fix nb_tensor stride check logic
The step responsible for checking whether a tensor has the right strides
to invoke a C++ routine was incorrect and always accepted C-style NumPy
arrays (because they leave the 'stride' field set to NULL).

The whole conversion logic will be more heavily tested in a subsequent
commit that adds a type caster for Eigen arrays.