Cleanup 8/10: added ``api<T>::operator~``, removed ``api<T>::operator!``
The previous implementation of ``api<T>::operator!`` was incorrect and
actually evaluated the Python ``~`` operator. This commit renames it.
The ``operator!`` is now missing---this will need some more attention in
the future, since it represents a special case (the ``not`` operator
always returns ``bool`` in Python).

This commit is part of a larger cleanup pass over API/implementation to
fix small inconsistencies that were found while documenting nanobind.