nanobind-config.cmake: Avoid top-level warnings in nanobind headers
If you're using nanobind_add_module within a project that has many warnings
enabled, the compiler may warn you that nanobind doesn't comply with the
current warning settings.

To prevent the warnings set by the top-level project from affecting nanobind,
we can add the nanobind include directory with the SYSTEM flag. This will mark
the directory as a system directory, which typically means the compiler will
not issue warnings for files included from this directory.