macOS linking: re-enable chained fixups using a linker response file
Following a high-bandwidth exchange with the Apple team in charge of the
macOS linker, this commit adjusts macOS extension linking so that it
retains the "chained fixups" optimization to produce smaller binaries.

This is done by informing the linker about every individual CPython (or
PyPy) API symbol that we expect to import dynamically at runtime. This
can actually be a relatively long list, and the commit uses a so-called
"linker response file" to specify them more efficiently.

Further technical details on chained linkups and the reasons for making
this change can be found in this issue tracker post:

https://github.com/python/cpython/issues/97524#issuecomment-1458855301