Disable implicit conversions in copy constructors
It's common to bind classes with a copy constructor, e.g.,

```cpp
nb::class_<T>(m, "T")
    .def(nb::init<const T &>());
```

A problematic issue can arise when `T` supports implicit conversions,
has a copy constructor, and when the user passes an argument to a
function expecting a `T` instance that isn't a perfect match for any
provided constructor overload.

nanobind then started a recursive implicit conversion spree that
eventually failed with a stack overflow + warning. Worse, that warning
was likely hidden, giving no indication of the existence of a problem
(see the previous commit for details and a fix).

This commit fixes the first part of the issue by disabling implicit
conversions in copy constructors.