Fix ``ndarray_import`` when the imported array has a dimension with size 0
This commit is needed to correctly handle the situation where nanobind
is invoked with a n-dimension array of, say, shape ``(100, 0, 20)``:

In PyTorch, a row-major tensor with this shape has strides ``(20, 20,
1)`` NumPy instead uses ``(0, 0, 0)``. The previous condition
`!(strides[i] == t.strides[i] || t.shape[i] <= 1)` was too strict to
handle all of these different behaviors. The solution is just to accept
any strides when given an empty array.