Merge supplement into type object
This commit updates the ``nb_type`` metaclass to create type objects
that directly include the supplement (if present). This involves a
dictionary of metaclasses (``nb_type_X``, where ``X`` is the size in
bytes) that are created on demand.

The commit also simplifies ``nb_type_new()`` by moving much of its
contents into a new function ``nb_type_from_metaclass()``, which simply
calls ``PyType_FromMetaclass()`` when it is available. Otherwise, it
emulates the behavior of ``PyType_FromMetaclass()``. This part has also
been simplified and no longer involves the creation of an intermediate
type.