Crunched internal assertion checks
Nanobind performs numerous checks to catch severe internal errors. It
prints a detailed error message and then calls ``abort()`` to stop the
process when such an unrecoverable failure is detected. The error
reporting unfortunately added quite a bit of code and data (~5-10KB) to
the binary, which is really just useful to developers working with
nanobind.

This commit switches the behavior so that checks are still done, but
detailed error reporting is no longer done in ``Release`` and
``MinSizeRel`` compilation modes. When an error is detected, the system
instead prints an error message asking the developer to recompile in
debug mode for more details.