Changes nb::tuple() to construct an empty tuple.
This is consistent with both `list` and `dict`. However, it looks like this might have been done this way for a reason so I'm posting this patch for comment.

Default initializing to nullptr differs from what pybind11 does, and I found this difference due to a segfault while porting some existing code.

Fixes #241