Fixed another sanitizer issue
``tests/test_intrusive.py::test04_subclass`` produced the below error in
the UB sanitizer. This is likely a fluke since the object in question is
being initialized (via placement new), not read. That said, it is easy
to suppress the warning, which is what this commit does.

```
/Users/wjakob/nanobind/include/nanobind/nb_class.h:312:22: runtime error: downcast of address 0x00010f167ed8 which does not point to an object of type 'Alias' (aka 'PyTest')
0x00010f167ed8: note: object has invalid vptr
 43 00 00 00  00 00 00 00 00 00 00 00  00 00 00 00 00 00 00 00  00 00 00 00 00 00 00 00  00 00 00 00
              ^~~~~~~~~~~~~~~~~~~~~~~
              invalid vptr
SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior /Users/wjakob/nanobind/include/nanobind/nb_class.h:312:22
```