Fix implicit conversion of ``nb::ndarray`` to contiguous layout
nanobind ndarrays provide the ``nb::c_contig`` and ``nb::f_contig``
annotations to specify that input arrays must be represented by
contiguous memory blocks in C or Fortran-style ordering. When this is
not the case, the nanobind will by default attempt an implicit
conversion.

This conversion previously failed in some cases: when no underlying
scalar type was specified, and when converting from PyTorch. Those
issues are addressed by this commit.

Fixes issue #278.