Officially support intrusive reference counting sample implementation
This commit turns what was previously just an example implementation of
intrusive reference counting for testing into an officially supported
feature that can now be accessed through 3 headers:

- ``include/nanobind/intrusive/counter.h``: Declarations
- ``include/nanobind/intrusive/counter.inl``: Implementation
- ``include/nanobind/intrusive/ref.h``: Reference counting RAII helper
  class and Python bindings

The documentation was updated to explain how to use these classes.

This commit does not create a dependence on this particular
implementation of intrusive reference counting---existing code should
remain unaffected.