Two-level hashing of ``std::type_info *`` via pointer & string hash (fixes #283)
Exchange of RTTI type information between separately compiled extension
libraries tends to be fragile: depending on the compiler and platform,
the C++ standard library may use one of two strategies to decide whether
two ``std::type_info`` instances are equal. The first is to perform a
string comparison of the mangled type name. When types are organized in
hash tables, a string hash is then also needed. This strategy yields the
expected result but can be rather inefficient.

Other platforms simply compare the pointer value and rely on a
poiner-based hashing scheme. This is far more efficient but requires
that the linker merges duplicate RTTI information from separate shared
libraries.

Unfortunately, this does not always work for the following reasons:

1. Python passes the ``RTLD_LOCAL``  flag to ``dlopen()`` when it loads
   shared libraries. If RTTI symbols are exported by a separate shared
   library, then things may still be fine. But if the Python extension
   is in charge of exporting RTTI symbols, there is problem.

2. It can generally be tricky to get the compiler to export RTTI symbols
   for non-polymorphic types.

3. Setting the right ``__attribute__ ((visibility("default")))`` flags
   in extension libraries is error-prone and a source of confusion for
   new users.

This commit changes nanobind to adopt both strategies at the same time.
Type queries first go through a fast pointer-based hash table followed
by a secondary string-based hash table. In the latter case, nanobind
also populates the faster pointer-based table with the missing
information so that the fast path eventually resolves all of the
queries.

This commit changes the internal representation of nanobind data
structures, hence the ABI version had to be incremented.