Reworked ``nb::raw_doc`` into a new interface named ``nb::signature``
The old ``nb::raw_doc`` API could previously be used to override the
docstring of an entire overload chain.

This has the downside that it cannot easily be split into per-overload
signatures and docstrings that are needed by the new stub generation
tool. This commit removes ``nb::raw_doc`` and replaces it with a new
annotation ``nb::signature`` that only overrides the function type
signature of a single overload, while keeping the docstring separate.

This is an API and ABI-breaking change.