Ability to override class binding signatures
This commit extends the ``nb::signature`` binding annotation so that it
can also be used to change how classes are rendered in stubs.

It also modifies and extends the syntax used to override function
signatures. With these changes, it is now possible to generate stubs
containing decorated functions and classes.