Add a `__dict__` attribute to nanobind::dynamic_attr() classes.
I found it surprising that class with dynamic attributes (i.e., it must
have a `__dict__` internally) didn't have a `o.__dict__` visible to
Python users. It seems that simply setting `tp_dictoffset` isn't enough
for `o.__dict__` to be present.

If the user didn't provide a `tp_getset`, populate it with `__dict__`
for classes with dynamic attributes.