Use a checked cast() in accessor cast operator.
Currently the following code pattern performs an unchecked cast:
```
nb::object o = ...;
nb::list l = o.attr("a");
```

The type of the attribute is not checked, which leads to surprising
errors when the type of the attribute is something else, e.g., a tuple.

It seems like a footgun to have implicit unchecked casts; casts should
either be implicit or unchecked but not both.