Fix degeneracy in ``inst_c2p`` and ``keep_alive`` hash tables (fixes #507)
Issue #507 revealed that creating large numbers of objects (potentially
with ``keep_alive`` records) caused the internal hash tables to grow to
accommodate these instances. So far, so good.

However, batch-deleting these objects then led to surprisingly poor
performance due to how the implementation of
``robin_map::erase(iterator)`` works. Erasing a hash entry causes the
iterator to walks to the next nonempty bucket. That becomes very
expensive when the table is nearly empty. In principle, a minimum load
factor can be specified, but the problem is that the ``erase(iterator)``
function does not re-hash the table even when its size goes below the
specified load factor.

The solution is to

1. specify a minimum load factor on these tables (previously, it was
   set to zero), and

2. use a variant of ``erase()``, specifically ``erase(key, hash)``
   that does not return an iterator.