Add documents explaining nanobind extension building using Bazel
This is designed to be a better-exposed doc on how to use nanobind
with Bazel, aimed at people who would like to get started with
nanobind but do not want to search around reading outside documentation.

Contains one usage guide document on how to declare the dependency,
how to import nanobind-bazel's APIs, how to use it in BUILD files,
and lastly how to build extensions with Bazel.

Adds an API reference directly behind the CMake API reference, marking it
as a third-party project. This API reference contains the build rules and
flags.

Mentions a few noteworthy special points of information, such as constraints
on the Bazel version, overriding versions with bzlmod APIs, and links to
other Python support projects in Bazel.