Recommend nanobind_ROOT over CMAKE_PREFIX_PATH
If I understand correctly, the goal of the recommended CMake code is to
ensure that the version of nanobind selected by `find_package` is
commensurate with the previously discovered Python environment.
`<PackageName>_ROOT` is a more targetted way to achieve this goal than
the global `CMAKE_PREFIX_PATH`.

This is functionally no change as `nanobind's` CMake config helper
is placed in an isolated directory. However, users may apply the CMake
code that applies `CMAKE_PREFIX_PATH` in other contexts and be surprised
when they modify the search order for *all* `find_package` commands.

https://cmake.org/cmake/help/latest/command/find_package.html#config-mode-search-procedure