Add nanobindConfigVersion.cmake and convert build to scikit-build-core (#618)
* Recommend nanobind_ROOT over CMAKE_PREFIX_PATH

If I understand correctly, the goal of the recommended CMake code is to
ensure that the version of nanobind selected by `find_package` is
commensurate with the previously discovered Python environment.
`<PackageName>_ROOT` is a more targetted way to achieve this goal than
the global `CMAKE_PREFIX_PATH`.

This is functionally no change as `nanobind's` CMake config helper
is placed in an isolated directory. However, users may apply the CMake
code that applies `CMAKE_PREFIX_PATH` in other contexts and be surprised
when they modify the search order for *all* `find_package` commands.

https://cmake.org/cmake/help/latest/command/find_package.html#config-mode-search-procedure

* Install nanobindConfigVersion.cmake

Define the nanobind version in CMake and install the version config file
so that downstream packages can require specific versions as needed.

* Use scikit-build-core

Switch from setup.py build to scikit-build-core. The existing
install rules in CMakeLists.txt already implemented the full installation
required.

* Remove additional unneeded build dependencies.

Co-authored-by: Henry Schreiner <HenrySchreinerIII@gmail.com>

* Exactly match files installed by `setup.py`.

CMake's directory glob was installing files that `setup.py` did not
install previously.

* Build py3-none-any wheels.

* fix: move rest of metadata over

Signed-off-by: Henry Schreiner <henryschreineriii@gmail.com>

* fix: add classifiers (cannonical location for License

Signed-off-by: Henry Schreiner <henryschreineriii@gmail.com>

* Determine nanobind version from nanobind.h.

This matches the behavior in the previous setup.py.

* Fix nanobind_VERSION CMake variable

CMake apparently requires `project-config-version.cmake` when using
`project-config`.cmake.

---------

Co-authored-by: Henry Schreiner <HenrySchreinerIII@gmail.com>