Separate const of ndarray from const of its data. (#491)
A static assertion ensures that the ``order`` and ``framework``
cannot be changed once set.  For example, a template argument to
``view()`` can only change one of these if it has not been set.
The scalar element type and the ``shape`` can be set multiple times,
and the rightmost one takes effect.  Thus, these can be changed by
specifying template arguments to ``view()``.
The read-only boolean, ``is_ro``, can be set either by specifying
``nb::ro`` or by specifying a const-qualified scalar element type.
It is sticky.  Once set, it cannot be cleared.  So, a read-only
view of writable data can be obtained, but not the reverse.