Move Bazel API defs into Python domain
This is a hack since there is no Bazel domain provider available on
PyPI, but it does seem to work since Starlark is a true subset of
Python, and we do not have any classes or module structure to take care
of.

Breaks the unlucky collision of `nanobind_extension` between the CMake
and Bazel docs that previously broke cross-referencing.