free-threading: GC adaptations
The garbage collector in free-threaded Python does not like it when the
reference count of an object temporarily increases while traversing the
object graph in ``tp_traverse``, and doing so introduces leaks.
Unfortunately, example implementations of ``tp_traverse`` in both
documentation and test suite fall into this trap and must be adapted.