Detect use of type vectorcall slot and use fallback implementation
Following the merge of https://github.com/python/cpython/pull/123332,
type slot 82 is reserved for type vector calls. This means that we can
safely intercept this value if provided to nanobind even on older
versions of Python where the feature isn't supported yet. In this case,
the metaclass-based implementation is used.