Custom dtype improvements (#742)
This commit generalizes the interface for registering custom dtypes with
the ``nb::ndarray`` class. In particular, the previous version had
hardcoded members like ``is_float``, ``is_int``, etc., which is not
general enough to express nuances like ``bfloat16`` or 8-bit FP types
with different numbers of exponent and mantissa bits.

The new interface just lets the developer populate a
``nb::dlpack::dtype`` record manually and provide a type name for use in
docstrings and error messages.